mc_engine/requirements.txt
ultralytics>=8.3.0
mediapipe==0.10.14
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.13.0
Pillow==10.3.0
pandas==2.2.2
tqdm==4.66.4
flask>=3.0.0
flask-cors>=4.0.0
redis>=5.0.0
python-dotenv>=1.0.0
